+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 
 
+## [0.5.4](https://github.com/Blobfolio/utc2k/releases/tag/v0.5.4) - TBD
+
+### Added
+
+* `Utc2k::cmp_date`
+* `Utc2k::cmp_time`
+
+
+
 ## [0.5.3](https://github.com/Blobfolio/utc2k/releases/tag/v0.5.3) - 2022-07-14
 
 ### Changed